attendance service changed

+++ symfony/plugins/orangehrmAttendancePlugin/lib/service/AttendanceService.php
@@ -87,7 +87,7 @@ class AttendanceService {
     /**
      * get saved Attendance configuration
      * @param $workflow, $state, $role, $action, $resultingState
-     * @return boolean 
+     * @return boolean
      */
     public function getSavedConfiguration($workflow, $state, $role, $action, $resultingState) {
 
@@ -95,9 +95,9 @@ class AttendanceService {
     }
 
     /**
-     * get Attendance record 
+     * get Attendance record
      * @param $employeeId, $date
-     * @return array of records 
+     * @return array of records
      */
     public function getAttendanceRecord($employeeId, $date) {
 
@@ -105,7 +105,7 @@ class AttendanceService {
     }
 
     /**
-     * delete Attendance record 
+     * delete Attendance record
      * @param $attendanceRecordId
      * @return boolean
      */
@@ -135,8 +135,8 @@ class AttendanceService {
     }
 
     /**
-     * Get Timezone Array 
-     * @param 
+     * Get Timezone Array
+     * @param
      * @return time zone values array
      */
     public function getTimezoneArray() {
@@ -225,8 +225,8 @@ class AttendanceService {
             return false;
         return $index[0];
     }
-    
-    
+
+
      /**
      * check For Punch In OverLapping Records when Editing
      * @param $punchInTime, $employeeId
@@ -235,8 +235,8 @@ class AttendanceService {
     public function checkForPunchInOverLappingRecordsWhenEditing($punchInTime, $employeeId,$recordId, $punchOut) {
         return $this->getAttendanceDao()->checkForPunchInOverLappingRecordsWhenEditing($punchInTime, $employeeId,$recordId, $punchOut);
     }
-    
-    
+
+
      /**
      * check For Punch out OverLapping Records when Editing
      * @param $punchInTime, $employeeId
@@ -245,7 +245,7 @@ class AttendanceService {
     public function checkForPunchOutOverLappingRecordsWhenEditing($punchIn, $punchOut, $employeeId,$recordId) {
         return $this->getAttendanceDao()->checkForPunchInOutOverLappingRecordsWhenEditing($punchIn, $punchOut, $employeeId, $recordId);
     }
-    
+
      /**
      * check For Punch out/in OverLapping Records when Editing
      * @param $punchInTime, $employeeId
@@ -254,15 +254,15 @@ class AttendanceService {
     public function checkForPunchInOutOverLappingRecordsWhenEditing($punchIn, $punchOut, $employeeId,$recordId) {
         return $this->getAttendanceDao()->checkForPunchInOutOverLappingRecordsWhenEditing($punchIn, $punchOut, $employeeId, $recordId);
     }
-    
+
     /**
      *
      * @param int $employeeId
      * @param string $employeementStatus
-     * @param int $subDivision    
+     * @param int $subDivision
      * @param date $dateFrom
      * @param date $dateTo
-     * @return array 
+     * @return array
      */
     public function searchAttendanceRecords($employeeId = null, $employeementStatus = null, $subDivision = null, $dateFrom = null , $dateTo = null ){
         return $this->getAttendanceDao()->searchAttendanceRecords($employeeId, $employeementStatus, $subDivision, $dateFrom, $dateTo );
@@ -280,8 +280,8 @@ class AttendanceService {
     }
 
     /**
-     * @param string $dateTime
-     * @param int $timeZoneOffset
+     * @param string $dateTime    example : "2021-01-31 19:55"
+     * @param int $timeZoneOffset  example : -9000
      * @return false|string
      */
     public function getCalculatedPunchInUtcTime(string $dateTime, int $timeZoneOffset)
@@ -306,11 +306,11 @@ class AttendanceService {
     }
 
     /**
-     * @param int $punchOutTimeOffset
+     * @param float $punchTimeOffset
      * @return float|int
      * @throws Exception
      */
-    public function getOriginDisplayTimeZoneOffset(int $punchTimeOffset)
+    public function getOriginDisplayTimeZoneOffset(float $punchTimeOffset)
     {
         $remoteDTZ = new DateTimeZone('UTC');
         $remoteDT = new DateTime("now", $remoteDTZ);
@@ -336,6 +336,5 @@ class AttendanceService {
         $zoneList = timezone_identifiers_list();
         return in_array($timeZone, $zoneList);
     }
-
 }
